refactor(tracing): narrow shared tool-attr helper, drop re-export
Addresses review feedback on PR #8800:

- Narrow getToolNameFromAttributes to an explicit-key lookup only. The
  permissive regex fallback it used to contain was previously local to
  trajectory extraction; sharing it with traceContext.deriveInsights and
  redteam traceFormatting would have silently matched attributes like
  tool.description or workflow.tool as tool names. Keep the regex
  fallback inline in extractToolName where it belongs.
- Drop the re-export of getToolNameFromAttributes from traceContext.ts
  per AGENTS.md ("Avoid re-exporting from files; import directly from
  the source module"). Update traceFormatting.ts to import from
  ./tracing/toolAttributes directly.
- Rename test/tracing/traceContext.test.ts to toolAttributes.test.ts to
  mirror the actual source module, and drop tests that exercised the
  now-removed regex fallback (still covered via trajectory tests).
- Reorder TOOL_ARGUMENT_ATTRIBUTE_KEYS to put generic tool.arguments
  before vendor-specific ai.toolCall.args, matching the ordering used
  for TOOL_NAME_ATTRIBUTE_KEYS.
